Increased the type size for the stop-after-n-instructions option to avoid too strict limitations

+++ lib/Core/Executor.cpp
@@ -290,7 +290,7 @@ namespace {
 		    clEnumValEnd),
 		  cl::ZeroOrMore);
 
-  cl::opt<unsigned int>
+  cl::opt<unsigned long long>
   StopAfterNInstructions("stop-after-n-instructions",
                          cl::desc("Stop execution after specified number of instructions (default=0 (off))"),
                          cl::init(0));
